+++ src/Faithlife.Tracing.AspNetCore/AspNetCoreTracing.cs
@@ -28,22 +28,22 @@ namespace Faithlife.Tracing.AspNetCore
 			app.Use(async (httpContext, next) =>
 			{
 				var headers = httpContext.Request.Headers;
-				var parentTrace = Tracer.ExtractTrace(x => headers[x]);
-				if (parentTrace != null || headers["X-B3-Flags"] == "1" || GetHashCode(Interlocked.Increment(ref s_requestCount)) % c_samplingPrecision < SamplingRate)
+				var parentSpan = Tracer.ExtractSpan(x => headers[x]);
+				if (parentSpan != null || headers["X-B3-Flags"] == "1" || GetHashCode(Interlocked.Increment(ref s_requestCount)) % c_samplingPrecision < SamplingRate)
 				{
-					var trace = Tracer.StartTrace(parentTrace, TraceKind.Server,
+					var span = Tracer.StartSpan(parentSpan, TraceSpanKind.Server,
 						new[]
 						{
-							(TraceTagNames.Service, serviceName),
-							(TraceTagNames.HttpHost, httpContext.Request.Host.ToString()),
-							(TraceTagNames.HttpMethod, httpContext.Request.Method),
-							(TraceTagNames.HttpUrl, httpContext.Request.GetEncodedUrl()),
-							(TraceTagNames.HttpPath, httpContext.Request.GetEncodedPathAndQuery()),
+							(SpanTagNames.Service, serviceName),
+							(SpanTagNames.HttpHost, httpContext.Request.Host.ToString()),
+							(SpanTagNames.HttpMethod, httpContext.Request.Method),
+							(SpanTagNames.HttpUrl, httpContext.Request.GetEncodedUrl()),
+							(SpanTagNames.HttpPath, httpContext.Request.GetEncodedPathAndQuery()),
 						});
 
 					httpContext.Items[c_serviceNameKey] = serviceName;
 
-					var provider = new RequestActionTraceProvider(trace);
+					var provider = new RequestActionTraceSpanProvider(span);
 					httpContext.Items[c_providerKey] = provider;
 
 					try
@@ -52,8 +52,8 @@ namespace Faithlife.Tracing.AspNetCore
 					}
 					finally
 					{
-						provider.CurrentTrace.SetTag(TraceTagNames.HttpStatusCode, httpContext.Response.StatusCode.ToString(CultureInfo.InvariantCulture));
-						provider.FinishRequestTrace();
+						provider.CurrentSpan.SetTag(SpanTagNames.HttpStatusCode, httpContext.Response.StatusCode.ToString(CultureInfo.InvariantCulture));
+						provider.FinishRequestSpan();
 					}
 				}
 				else
@@ -69,7 +69,7 @@ namespace Faithlife.Tracing.AspNetCore
 			return builder;
 		}
 
-		public static ITraceProvider GetProvider(HttpContext context) => GetRequestActionTraceProvider(context) ?? NullTraceProvider.Instance;
+		public static ITraceSpanProvider GetProvider(HttpContext context) => GetRequestActionTraceSpanProvider(context) ?? NullTraceSpanProvider.Instance;
 
 		internal static ITracer Tracer { get; set; }
 
@@ -77,7 +77,7 @@ namespace Faithlife.Tracing.AspNetCore
 
 		internal static string GetServiceName(HttpContext context) => (string) context?.Items[c_serviceNameKey];
 
-		internal static RequestActionTraceProvider GetRequestActionTraceProvider(HttpContext context) => (RequestActionTraceProvider) context?.Items[c_providerKey];
+		internal static RequestActionTraceSpanProvider GetRequestActionTraceSpanProvider(HttpContext context) => (RequestActionTraceSpanProvider) context?.Items[c_providerKey];
 
 		// From http://burtleburtle.net/bob/hash/integer.html
 		private static uint GetHashCode(int value)