+++ src/Faithlife.Tracing.Zipkin/ZipkinTracer.cs
@@ -34,33 +34,33 @@ namespace Faithlife.Tracing.Zipkin
 			TraceManager.Start(new NullLogger());
 		}
 
-		public ITrace ExtractTrace(Func<string, string> extractKey)
+		public ITraceSpan ExtractSpan(Func<string, string> extractKey)
 		{
 			var traceContext = s_w3cExtractor.Extract(extractKey) ?? s_b3Extractor.Extract(extractKey);
-			return traceContext == null ? null : new ZipkinTrace(this, Trace.CreateFromId(traceContext), TraceKind.Server);
+			return traceContext == null ? null : new ZipkinTraceSpan(this, Trace.CreateFromId(traceContext), TraceSpanKind.Server);
 		}
 
-		public void InjectTrace(ITrace trace, Action<string, string> inject)
+		public void InjectSpan(ITraceSpan traceSpan, Action<string, string> inject)
 		{
-			s_injector.Inject(((ZipkinTrace) trace).WrappedTrace.CurrentSpan, inject);
+			s_injector.Inject(((ZipkinTraceSpan) traceSpan).WrappedTrace.CurrentSpan, inject);
 		}
 
-		public ITrace StartTrace(ITrace parent, TraceKind kind, IEnumerable<(string Name, string Value)> tags)
+		public ITraceSpan StartSpan(ITraceSpan parent, TraceSpanKind kind, IEnumerable<(string Name, string Value)> tags)
 		{
-			var trace = (parent as ZipkinTrace)?.WrappedTrace.Child() ?? Trace.Create();
-			if (kind == TraceKind.Client)
+			var trace = (parent as ZipkinTraceSpan)?.WrappedTrace.Child() ?? Trace.Create();
+			if (kind == TraceSpanKind.Client)
 				trace.Record(Annotations.ClientSend());
-			else if (kind == TraceKind.Server)
+			else if (kind == TraceSpanKind.Server)
 				trace.Record(Annotations.ServerRecv());
-			var zipkinTrace = new ZipkinTrace(this, trace, kind);
+			var zipkinTraceSpan = new ZipkinTraceSpan(this, trace, kind);
 			foreach (var (name, value) in tags)
 			{
-				if (kind == TraceKind.Local && name == TraceTagNames.Operation)
+				if (kind == TraceSpanKind.Local && name == SpanTagNames.Operation)
 					trace.Record(Annotations.LocalOperationStart(value));
 				else
-					zipkinTrace.SetTag(name, value);
+					zipkinTraceSpan.SetTag(name, value);
 			}
-			return zipkinTrace;
+			return zipkinTraceSpan;
 		}
 
 		static readonly Getter<Func<string, string>, string> s_callbackGetter = (f, k) => f(k);