+++ src/Faithlife.Tracing/ITracer.cs
@@ -13,23 +13,23 @@ namespace Faithlife.Tracing
 		/// Creates a new span (that is optionally a child of an existing span).
 		/// </summary>
 		/// <param name="parent">The parent span of the new span. This parameter may be <c>null</c> to create a new trace.</param>
-		/// <param name="kind">The <see cref="TraceKind"/>.</param>
-		/// <param name="tags">A sequence of span tags; see <see cref="ITrace.SetTag"/>.</param>
-		/// <returns>A new <see cref="ITrace"/>.</returns>
-		ITrace StartTrace(ITrace parent, TraceKind kind, IEnumerable<(string Name, string Value)> tags);
+		/// <param name="kind">The <see cref="TraceSpanKind"/>.</param>
+		/// <param name="tags">A sequence of span tags; see <see cref="ITraceSpan.SetTag"/>.</param>
+		/// <returns>A new <see cref="ITraceSpan"/>.</returns>
+		ITraceSpan StartSpan(ITraceSpan parent, TraceSpanKind kind, IEnumerable<(string Name, string Value)> tags);
 
 		/// <summary>
-		/// Extracts a trace from a string-to-string map.
+		/// Extracts a span from a string-to-string map.
 		/// </summary>
 		/// <param name="extractKey">A <see cref="Func{T,TResult}"/> that looks up a value given a string key.</param>
-		/// <returns>A new <see cref="ITrace"/> if one could be extracted successfully, or <c>null</c>.</returns>
-		ITrace ExtractTrace(Func<string, string> extractKey);
+		/// <returns>A new <see cref="ITraceSpan"/> if one could be extracted successfully, or <c>null</c>.</returns>
+		ITraceSpan ExtractSpan(Func<string, string> extractKey);
 
 		/// <summary>
-		/// Injects a trace into a string-to-string map.
+		/// Injects a span into a string-to-string map.
 		/// </summary>
-		/// <param name="trace">The <see cref="ITrace"/> to inject.</param>
+		/// <param name="traceSpan">The <see cref="ITraceSpan"/> to inject.</param>
 		/// <param name="inject">A <see cref="Action{T1,T2}"/> that sets a key to a specified value.</param>
-		void InjectTrace(ITrace trace, Action<string, string> inject);
+		void InjectSpan(ITraceSpan traceSpan, Action<string, string> inject);
 	}
 }